conns: remove region assignment for some global services (e.g. route53)
It's not clear whether these global services require a specific region
to be assigned. For example, IAM isn't in this list.
The idea probably was to specify the primary region of the partition,
regardless of what is specified in the tf configuration.

We can't just use ru-msk for c2 partition here, because this partition
is the same for every installation.

Probably it is possible to generate clients for these services now
(see `SkipClientGenerate`).